% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_replace.R
\name{line_replace}
\alias{line_replace}
\title{Replace lines within char-vector}
\usage{
line_replace(txt, parameter, pattern)
}
\arguments{
\item{txt}{File input}

\item{parameter}{Content to replace}

\item{pattern}{Content to look for}
}
\value{
txt, with line containing \var{pattern} replaced with \var{parameter}
}
\description{
Replace lines within char-vector
}
