% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mcmc.R
\name{read_mcmc}
\alias{read_mcmc}
\title{read_mcmc}
\usage{
read_mcmc(file, cn = "value")
}
\arguments{
\item{file}{A character vector, the prefix to the .mcmc file to read}

\item{cn}{Name of the variable}
}
\value{
a tibble
}
\description{
read_mcmc
}
