% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prognosis.R
\name{set_prognosis}
\alias{set_prognosis}
\title{set_prognosis}
\usage{
set_prognosis(infile, outfile, fbar, hr, w.rho, w.cv, rec.rho, ass.cv,
  ass.rho, ass.bias, Btrigger, MaxChange, mW.years, LastYearsTacRatio)
}
\arguments{
\item{LastYearsTacRatio}{}
}
\description{
set_prognosis
}
