% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{summarise_results}
\alias{summarise_results}
\title{Title}
\usage{
summarise_results(d, n = 5)
}
\arguments{
\item{d}{A dataframe, normally that returned by gather_results}

\item{n}{A value, indicating the number of "last years" used in the summarisation}
}
\value{
A tibble
}
\description{
Title
}
